allow null audio source
currently, null audio source is only for where the original video source does
not have an audio source, such as a Raspberry PiCamera or scientific camera.

For sources that already have an audio stream, errors will result and it will
not work. Source mapping would be the fix for this.